+++ CalculoIRRF/Model/Dependente.cs
@@ -6,5 +6,5 @@ public class Dependente
 {
     public int Id { get; set; }
     public DateTime Competencia { get; set; }
-    public decimal Valor { get; set; }
+    public double Valor { get; set; }
 }
