+++ CalculoIRRF/Model/Inss.cs
@@ -7,6 +7,6 @@ public class Inss
     public int Id { get; set; }
     public DateTime Competencia { get; set; }
     public int Faixa { get; set; }
-    public decimal Valor { get; set; }
-    public decimal Porcentagem { get; set; }
+    public double Valor { get; set; }
+    public double Porcentagem { get; set; }
 }
