+++ CalculoIRRF/Model/Irrf.cs
@@ -7,8 +7,8 @@ public class Irrf
     public int Id { get; set; }
     public DateTime Competencia { get; set; }
     public int Faixa { get; set; }
-    public decimal Valor { get; set; }
-    public decimal Porcentagem { get; set; }
-    public decimal Deducao { get; set; }
+    public double Valor { get; set; }
+    public double Porcentagem { get; set; }
+    public double Deducao { get; set; }
 
 }
