+++ CalculoIRRF/Repository/Interface/IInssRepository.cs
@@ -10,11 +10,11 @@ public interface IInssRepository
     Task<IEnumerable<Inss>> GetAll();
     Task<IEnumerable<Inss>> GetByCompetence(DateTime competence);
     Task<Inss> GetById(int id);
-    Task<int> Range(decimal baseInss, DateTime competence);
+    Task<int> Range(double baseInss, DateTime competence);
     Task<int> LastRange(DateTime competence);
-    Task<decimal> Percent(int range, DateTime competence);
-    Task<decimal> Value(int range, DateTime competence);
-    Task<decimal> Roof(DateTime competence);
+    Task<double> Percent(int range, DateTime competence);
+    Task<double> Value(int range, DateTime competence);
+    Task<double> Roof(DateTime competence);
     Task<Inss> Create(Inss inss);
     Task<Inss> Update(Inss inss);
     Task<Inss> Delete(int id);
